+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Peer to Peer mode with CLI and ENV options to enable, added for DHT node using the OpenDHT library
 - RESTful API numbers resource (`/numbers`) now only returns numbers like +441234567890 or 441234567890
 - RESTful API number resource (`/numbers/{number}`) now only accepts +441234567890 or 441234567890 formats